Error en manejo de errores de delete solucionado

+++ API_REST/src/logic/Delete.java
@@ -12,16 +12,22 @@ public class Delete {
 
 	public static void doDelete(HttpServletRequest request) throws IllegalArgumentException{
 		Long indice = (long) Integer.parseInt(request.getParameter("indice"));
+		
 
 		Session session = HibernateUtil.getSession();
 		Censo c = session.get(Censo.class, indice);
-
+if(c ==null) {
+	throw new IllegalArgumentException();
+}else {
 		session.beginTransaction();
 		session.delete(c);
 		session.getTransaction().commit();
-
-		session.close();
+		
+		
 
 		System.out.println("El censo seleccionado ha sido eliminado: " + c);
+		session.close();
+}
+
 }
 }
